refactor: applied the SOLID
